feat(rabbitmq) use execute_request for all endpoints, refactoring
