+++ app/endpoints/audio.py
@@ -62,8 +62,8 @@ async def audio_transcriptions(
         return JSONResponse(content=AudioTranscription(**response.json()).model_dump(),
                             status_code=response.status_code)
 
-    model = await global_context.model_registry(model=model)
-    return await model.safe_client_access(
+    return await global_context.model_registry.execute_request(
+        router_id=model,
         endpoint=ENDPOINT__AUDIO_TRANSCRIPTIONS,
         handler=handler
     )
