+++ app/endpoints/completions.py
@@ -19,8 +19,8 @@ async def completions(request: Request, body: CompletionRequest) -> JSONResponse
         response = await client.forward_request(method="POST", json=body.model_dump())
         return JSONResponse(content=Completions(**response.json()).model_dump(), status_code=response.status_code)
 
-    model = await global_context.model_registry(model=body.model)
-    return await model.safe_client_access(
+    return await global_context.model_registry.execute_request(
+        router_id=body.model,
         endpoint=ENDPOINT__COMPLETIONS,
         handler=handler
     )
