+++ app/endpoints/embeddings.py
@@ -19,8 +19,8 @@ async def embeddings(request: Request, body: EmbeddingsRequest) -> JSONResponse:
         response = await client.forward_request(method="POST", json=body.model_dump())
         return JSONResponse(content=Embeddings(**response.json()).model_dump(), status_code=response.status_code)
 
-    model = await global_context.model_registry(model=body.model)
-    return await model.safe_client_access(
+    return await global_context.model_registry.execute_request(
+        router_id=body.model,
         endpoint=ENDPOINT__EMBEDDINGS,
         handler=handler
     )
