+++ app/endpoints/rerank.py
@@ -18,8 +18,8 @@ async def rerank(request: Request, body: RerankRequest) -> JSONResponse:
         response = await client.forward_request(method="POST", json=body.model_dump())
         return JSONResponse(content=Reranks(**response.json()).model_dump(), status_code=response.status_code)
 
-    model = await global_context.model_registry(model=body.model)
-    return await model.safe_client_access(
+    return await global_context.model_registry.execute_request(
+        router_id=body.model,
         endpoint=ENDPOINT__RERANK,
         handler=handler
     )
